Add files via upload
Not every file in this Repo is actually used in the MVP. However, I have included those unused files to demonstrate my thought and development progress. This was perhaps the most challenging project I've ever worked on, and was very stressful, almost sickening, and made me panic quite a few times, but thank God, I got it done.